Added SG13_via_lib to autorun, added module with __init__
