+++ ihp-sg13g2/libs.tech/klayout/python/sg13g2_native_pcell_lib/__init__.py
@@ -1,0 +1,31 @@
+########################################################################
+#
+# Copyright 2024 IHP PDK Authors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    https://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+########################################################################
+
+import pya
+from .via_pcell import ViaPCell
+
+
+class SG13G2_ViaLib(pya.Library):
+    def __init__(self):
+        self.description = "SG13G2 Via Stack Library"
+        self.layout().register_pcell("ViaStack", ViaPCell())
+        self.register("SG13_via_lib")
+        self.technology = 'sg13g2'
+
+
+SG13G2_ViaLib()
