+++ ihp-sg13g2/libs.tech/klayout/tech/pymacros/autorun.lym
@@ -37,5 +37,6 @@ import os
 sys.path.append(os.path.join(os.path.dirname(__file__), '..', '..', 'python/'))
 sys.path.append(os.path.join(os.path.dirname(__file__), '..', '..', 'python/pycell4klayout-api/source/python/'))
 import sg13g2_pycell_lib
+import sg13g2_native_pcell_lib
 </text>
 </klayout-macro>
